+++ dataland-frontend/src/router/index.ts
@@ -18,8 +18,7 @@ const BulkDataRequest = (): Promise<RouteComponent> => import("@/components/page
 const SingleDataRequest = (): Promise<RouteComponent> => import("@/components/pages/SingleDataRequest.vue");
 const ViewFrameworkData = (): Promise<RouteComponent> => import("@/components/pages/ViewFrameworkData.vue");
 const DatasetOverview = (): Promise<RouteComponent> => import("@/components/pages/DatasetOverview.vue");
-const RequestedDatasetsPage = (): Promise<RouteComponent> =>
-  import("@/components/pages/MyRequestedDatasetsOverview.vue");
+const RequestedDatasetsPage = (): Promise<RouteComponent> => import("@/components/pages/MyDataRequestsOverview.vue");
 const UploadFormWrapper = (): Promise<RouteComponent> => import("@/components/pages/UploadFormWrapper.vue");
 const ChooseCompanyForFrameworkDataUpload = (): Promise<RouteComponent> =>
   import("@/components/pages/ChooseCompanyForFrameworkDataUpload.vue");
